# packages/contracts/openapi/patchpilot.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PatchPilot Control Plane API",
    "version": "patchpilot.mvp.v1",
    "description": "Generated from @patchpilot/contracts artifacts used by API, Web, and Worker packages."
  },
  "paths": {
...
    "/api/prds/:id/audit-export": {
      "get": {
        "operationId": "exportPrdAuditPackage",
        "tags": [
          "Control Plane"
        ],
        "summary": "GET /api/prds/:id/audit-export",
        "security": [
          {
            "PatchPilotUser": [],
            "PatchPilotRole": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditExportPackage"
                }
              }
            }
          },
          "401": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/requirements": {
      "post": {
        "operationId": "createRequirement",
        "tags": [
          "Control Plane"
        ],
        "summary": "POST /api/requirements",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRequirementRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Requirement"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/requirements/:id/clarification-answer": {
      "post": {
        "operationId": "answerClarification",
        "tags": [
          "Control Plane"
        ],
        "summary": "POST /api/requirements/:id/clarification-answer",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClarificationAnswerRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequirementPrdBundle"
                }
              }
            }
          }
        }
      }
    },
    "/api/requirements/:id/clarification-turn": {
      "post": {
        "operationId": "addClarificationTurn",
        "tags": [
          "Control Plane"
        ],
        "summary": "POST /api/requirements/:id/clarification-turn",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClarificationTurnRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClarificationTurnResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/prds/:id/approve": {
      "post": {
        "operationId": "approvePrd",
        "tags": [
          "Control Plane"
        ],
        "summary": "POST /api/prds/:id/approve",
        "security": [
          {
            "PatchPilotUser": [],
            "PatchPilotRole": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovePrdResponse"
                }
              }
            }
          },
          "401": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/prds/:id/start-team": {
      "post": {
        "operationId": "startTeam",
        "tags": [
          "Control Plane"
        ],
        "summary": "POST /api/prds/:id/start-team",
        "security": [
          {
            "PatchPilotUser": [],
            "PatchPilotRole": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartRunRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartTeamResponse"
                }
              }
            }
          },
          "401": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/prds/:id/acceptance": {
      "post": {
        "operationId": "acceptTeam",
        "tags": [
          "Control Plane"
        ],
        "summary": "POST /api/prds/:id/acceptance",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptanceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamAcceptanceResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/approvals": {
      "post": {
        "operationId": "createApproval",
        "tags": [
          "Control Plane"
        ],
        "summary": "POST /api/approvals",
        "security": [
          {
            "PatchPilotUser": [],
            "PatchPilotRole": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApprovalRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalRecord"
                }
              }
            }
          },
          "401": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/approvals/:id/approve": {
      "post": {
        "operationId": "approveApproval",
        "tags": [
          "Control Plane"
        ],
        "summary": "POST /api/approvals/:id/approve",
        "security": [
          {
            "PatchPilotUser": [],
            "PatchPilotRole": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalDecisionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalRecord"
                }
              }
            }
          },
          "401": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/approvals/:id/deny": {
      "post": {
        "operationId": "denyApproval",
        "tags": [
          "Control Plane"
        ],
        "summary": "POST /api/approvals/:id/deny",
        "security": [
          {
            "PatchPilotUser": [],
            "PatchPilotRole": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalDecisionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalRecord"
                }
              }
            }
          },
          "401": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/work-items/:id/start": {
      "post": {
        "operationId": "startWorkItem",
        "tags": [
          "Work Items"
        ],
        "summary": "POST /api/work-items/:id/start",
        "security": [
          {
            "PatchPilotUser": [],
            "PatchPilotRole": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartRunRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRun"
                }
              }
            }
          },
          "401": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/work-items/:id/claim": {
      "post": {
        "operationId": "claimWorkItem",
        "tags": [
          "Work Items"
        ],
        "summary": "POST /api/work-items/:id/claim",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClaimWorkItemRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClaimWorkItemResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/work-items/:id/release": {
      "post": {
        "operationId": "releaseWorkItem",
        "tags": [
          "Work Items"
        ],
        "summary": "POST /api/work-items/:id/release",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReleaseWorkItemRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReleaseWorkItemResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/bugs": {
      "post": {
        "operationId": "createBug",
        "tags": [
          "Bugs"
        ],
        "summary": "POST /api/bugs",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBugRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateBugResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/integrations/issues/link": {
      "post": {
        "operationId": "linkExternalIssue",
        "tags": [
          "Integrations"
        ],
        "summary": "POST /api/integrations/issues/link",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExternalIssueLinkRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalIssueSyncResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/integrations/issues/status": {
      "post": {
        "operationId": "updateExternalIssueStatus",
        "tags": [
          "Integrations"
        ],
        "summary": "POST /api/integrations/issues/status",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExternalIssueStatusUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalIssueSyncResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/integrations/github/repositories/select": {
      "post": {
        "operationId": "selectGitHubAppRepository",
        "tags": [
          "Integrations"
        ],
        "summary": "POST /api/integrations/github/repositories/select",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelectGitHubRepositoryRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GitHubAppRepositorySelectionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/integrations/github/webhook": {
      "post": {
        "operationId": "ingestGitHubAppWebhook",
        "tags": [
          "Integrations"
        ],
        "summary": "POST /api/integrations/github/webhook",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GitHubWebhookRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GitHubWebhookResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/acceptance/:runId": {
      "post": {
        "operationId": "acceptRun",
        "tags": [
          "Control Plane"
        ],
        "summary": "POST /api/acceptance/:runId",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptanceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcceptanceDecision"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "PatchPilotUser": {
        "type": "apiKey",
        "in": "header",
        "name": "x-patchpilot-user",
        "description": "PatchPilot authenticated actor id for protected control-plane actions."
      },
      "PatchPilotRole": {
        "type": "apiKey",
        "in": "header",
        "name": "x-patchpilot-role",
        "description": "PatchPilot actor role. Allowed values: submitter, maintainer, reviewer, admin."
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "requiredHeaders": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "AuditExportPackage": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "manifest",
          "verification",
          "records",
          "artifactManifest",
          "auditEvents",
          "auditEventsJsonl",
          "readme"
        ],
        "properties": {
          "manifest": {
            "$ref": "#/components/schemas/AuditExportManifest"
          },
          "verification": {
            "$ref": "#/components/schemas/AuditExportVerification"
          },
          "records": {
            "$ref": "#/components/schemas/AuditExportRecords"
          },
          "artifactManifest": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuditExportArtifactManifestEntry"
            }
          },
          "auditEvents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuditEvent"
            }
          },
          "auditEventsJsonl": {
            "type": "string"
          },
          "readme": {
            "type": "string"
          }
        }
      },
...
      "CreateRequirementRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "rawInput",
          "template"
        ],
        "properties": {
          "rawInput": {
            "type": "string",
            "minLength": 3
          },
          "template": {
            "type": "string",
            "enum": [
              "feature",
              "bug",
              "ui",
              "document"
            ]
          },
          "artifactReferences": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IntakeArtifactReferenceInput"
            }
          }
        }
      },
...
      "Requirement": {
        "type": "object",
        "additionalProperties": true,
        "required": [
          "id",
          "title",
          "rawInput",
          "template",
          "status",
          "simpleSummary",
          "artifactReferences",
          "clarificationQuestions",
          "clarificationTurns",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "minLength": 1
          },
          "title": {
            "type": "string"
          },
          "rawInput": {
            "type": "string"
          },
          "template": {
            "type": "string",
            "enum": [
              "feature",
              "bug",
              "ui",
              "document"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "submitted",
              "clarifying",
              "prd_draft",
              "approved",
              "rejected"
            ]
          },
          "simpleSummary": {
            "type": "string"
          },
          "artifactReferences": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IntakeArtifactReference"
            }
          },
          "clarificationQuestions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ClarificationQuestion"
            }
          },
          "clarificationTurns": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ClarificationTurn"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  }
}